Fix: There was a problem with displaying some kind of tortoise's pictures.

+++ src/Parser/TempAstNode.cpp
@@ -16,37 +16,37 @@ void TempAstNode::execute(Executer* executer)
 {
 	if(operation == "FD")
 	{
-		window->addOperation(std::pair<int, int>(0,children[0]->calc(executer)));
+		window->addOperation(std::pair<int, double>(0,children[0]->calc(executer)));
 	}
 	else if(operation == "BK")
 	{
-		window->addOperation(std::pair<int, int>(1,children[0]->calc(executer)));
+		window->addOperation(std::pair<int, double>(1,children[0]->calc(executer)));
 	}
 	else if(operation == "RT")
 	{
-		window->addOperation(std::pair<int, int>(2,children[0]->calc(executer)));
+		window->addOperation(std::pair<int, double>(2,children[0]->calc(executer)));
 	}
 	else if(operation == "LT")
 	{
-		window->addOperation(std::pair<int, int>(3,children[0]->calc(executer))); 
+		window->addOperation(std::pair<int, double>(3,children[0]->calc(executer))); 
 	}
 	else if(operation == "PU")
 	{
-		window->addOperation(std::pair<int, int>(4,0));
+		window->addOperation(std::pair<int, double>(4,0));
 	}
 	else if(operation == "PD")
 	{
-		window->addOperation(std::pair<int, int>(5,0));
+		window->addOperation(std::pair<int, double>(5,0));
 	}
 	else if(operation == "CS")
 	{
-		window->addOperation(std::pair<int, int>(9,0));
+		window->addOperation(std::pair<int, double>(9,0));
 	}
 	else if (operation == "SETPC")
 	{ 
-		window->addOperation(std::pair<int, int>(6,children[0]->calc(executer)));
-		window->addOperation(std::pair<int, int>(7,children[1]->calc(executer)));
-		window->addOperation(std::pair<int, int>(8,children[2]->calc(executer)));
+		window->addOperation(std::pair<int, double>(6,children[0]->calc(executer)));
+		window->addOperation(std::pair<int, double>(7,children[1]->calc(executer)));
+		window->addOperation(std::pair<int, double>(8,children[2]->calc(executer)));
 	}
 	
 }
